Rename level-details module and cleanup imports JW8-8804
