+++ src/loader/level-details.ts
